authentication

+++ app/Http/Controllers/CategoryController.php
@@ -2,41 +2,50 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\CategoryRequest;
+use App\Http\Requests\StoreCategoryRequest;
+use App\Http\Requests\UpdateCategoryRequest;
 use App\Http\Resources\CategoryResource;
 use App\Models\Category;
 
 class CategoryController extends Controller
 {
     public function index() {
-        return CategoryResource::collection(Category::orderByDesc('updated_at')->get());
+        $categories = Category::with('products')->orderByDesc('updated_at')->get();
+
+        return CategoryResource::collection($categories);
     }
 
     public function show(Category $category) {
-        return new CategoryResource($category);
+        return new CategoryResource($category->load('products'));
     }
 
-    public function store(CategoryRequest $request) {
-        $category = Category::create(['name' => $request->name]);
+    public function store(StoreCategoryRequest $request) {
+        $category = new Category(['name' => $request->name]);
+        $category->user()->associate($request->user());
+        $category->save();
 
         return [
-            'status' => 'Category was created successfully',
+            'message' => 'Category was created successfully',
             'category' => new CategoryResource($category)
         ];
     }
 
-    public function update(CategoryRequest $request, Category $category) {
-        $category->update(['name' => $request->name]);
+    public function update(UpdateCategoryRequest $request, Category $category) {
+        $this->authorize('update', $category);
+
+        $category->update(['name' => $request->name ?? $category->name]);
 
         return [
-            'status' => 'Category was updated successfully',
-            'category' => new CategoryResource($category)
+            'message' => 'Category was updated successfully',
+            'category' => new CategoryResource($category->load('products'))
         ];
     }
 
     public function destroy(Category $category) {
+        $this->authorize('delete', $category);
+
         $category->delete();
 
-        return ['status' => 'Category was deleted successfully'];
+        return ['message' => 'Category was deleted successfully'];
     }
 }
